requirements.txt
web3>=6
python-dotenv>=1
pytest>=8
hypothesis>=6  
ipykernel
